+++ package.json
@@ -8,9 +8,9 @@
   "scripts": {
     "lint": "eslint",
     "lint:fix": "eslint --fix",
-    "test": "npm run test:unit",
+    "test": "npm run test:unit && npm run test:typescript",
     "test:unit": "c8 --100 node --test",
-    "test:typescript": "tsd"
+    "test:typescript": "tstyche"
   },
   "repository": {
     "type": "git",
@@ -67,6 +67,6 @@
     "neostandard": "^0.13.0",
     "proxyquire": "^2.1.3",
     "sinon": "^21.0.0",
-    "tsd": "^0.33.0"
+    "tstyche": "^7.0.0"
   }
 }
